refactor(beastiary): extract view logic into child components
Moves creature detail panel, grid cards, and table rendering out of
BeastiaryView into dedicated CreatureDetail, CreatureGrid, and
CreaturesTable components. Each component now owns its own display
logic and collection state access.
